:bug: [Input] Boundary Conditions are Optional (#14)
Problem:
- The `boundary_condition` member function for the `Input` class
always returns a boundary condition, where the default is a wall if no
boundary condition was specified.  This is not compatible with
periodic boundaries or pressure drop boundaries where there are two
boundaries linked by the condition

Solution:
- Return an optional value from `boundary_condition` and let the
receiver decide whether it is a wall by default for linked to another
boundary.